+++ Package.swift
@@ -4,6 +4,7 @@ import PackageDescription
 
 let package = Package(
     name: "SwiftUIKit",
+    defaultLocalization: "en",
     platforms: [
         .iOS(.v15),
         .tvOS(.v15),
@@ -21,7 +22,8 @@ let package = Package(
     targets: [
         .target(
             name: "SwiftUIKit",
-            dependencies: []
+            dependencies: [],
+            resources: [.process("Resources")]
         ),
         .testTarget(
             name: "SwiftUIKitTests",