+++ Sources/SwiftUIKit/Lists/ListHeader.swift
@@ -84,25 +84,29 @@ public extension Image {
 
 #Preview {
     
-    VStack {
+    func item() -> some View {
+        Text("Preview.Item", bundle: .module)
+    }
+    
+    return VStack {
         List {
             ListHeader {
                 Color.red.frame(square: 150)
             }
             Section {
-                Text("Item")
-                Text("Item")
-                Text("Item")
-                Text("Item")
+                item()
+                item()
+                item()
+                item()
             }
         }
         List {
             Image(systemName: "face.smiling").listHeader(height: 75)
             Section {
-                Text("Item")
-                Text("Item")
-                Text("Item")
-                Text("Item")
+                item()
+                item()
+                item()
+                item()
             }
         }
     }